move ExtractCOSEPayload from enclaveapi.parsing to validation
